# [pkg/pdatautil] Optimize the hashing function for `pcommon.Map`